Version 0.0.8 the Pursuit instance has a local scope and can check for instanceOf
The method directory has access to variables in a scope that is private to the Pursuit instance. This is done by changing the function scope of the returned Pursuit function and the directory functions upon compilationtime --thus it is possible to store and receive data in the local scope.

Look in the implementation of the 'instanceOf'-directory function for an example of setting and getting data from the local scope.